Playing around with limbs and such

+++ Code/print_cells_csv.py
@@ -3,30 +3,31 @@ import matplotlib.pyplot as plt
 import os
 
 def print_cells_csv():
-    layers = [
-        'L23',
-        'L4',
-        'L6a'
-    ]
+    read_path = "cell_fire_locations.csv"
 
-    cell_type = 'active'
+    if os.path.exists(read_path):
+        layers = [
+            'L23',
+            'L4',
+            'L6a'
+        ]
 
-    read_path = "cell_fire.csv"
+        cell_type = 'active'
 
-    df = pd.read_csv(read_path)
+        df = pd.read_csv(read_path)
 
-    selected_columns = ['layer', 'thought_count', 'cell', 'column', 'x', 'y', 'head_direction','linear_speed','angular_velocity']
+        selected_columns = ['layer', 'thought_count', 'cell', 'column', 'x', 'y', 'head_direction','linear_speed','angular_velocity']
 
-    for l in range(len(layers)):
-        write_path =   f'C:\Environments\HTMAI Output\\{layers[l]}_{cell_type}'
-        path_exists = os.path.exists(write_path)
+        for l in range(len(layers)):
+            write_path =   f'C:\Environments\HTMAI Output\\{layers[l]}_{cell_type}'
+            path_exists = os.path.exists(write_path)
 
-        if not path_exists:
-            os.makedirs(write_path)
+            if not path_exists:
+                os.makedirs(write_path)
 
-        layer_data = df.loc[df['layer'] == f'{layers[l]}_{cell_type}']
+            layer_data = df.loc[df['layer'] == f'{layers[l]}_{cell_type}']
 
-        for i in range(256):
-            j = i*32
-            if not layer_data.loc[layer_data['cell'] == j].empty:
-                layer_data.loc[(layer_data['cell'] == j)].to_csv(f'{write_path}\cell{j}.csv')
+            for i in range(256):
+                j = i*32
+                if not layer_data.loc[layer_data['cell'] == j].empty:
+                    layer_data.loc[(layer_data['cell'] == j)].to_csv(f'{write_path}\cell{j}.csv')
